backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0
httpx>=0.26.0
sse-starlette>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Optional: ChromaDB for RAG (install with: pip install socraticgemma[rag])
# chromadb>=0.4.0
